refactor: change TestNovaAction namespace
